Added initial files.
A Media Player Controller using Hand Gestures has been developed. 
The tech stack of the project includes the concepts of Image Processing using OpenCV and Python.
The VLC Media Player is controlled with the help of hand gestures, without having to touch the keyboard. 
The controls that have been added are: play, pause, volume up
and volume down.